Lab-I-Dimensionality-Reduction/app-MNIST-prediction/requirements.txt
matplotlib==3.7.1
scikit-learn==1.2.2
#dimensionality_reductions_jmsv==0.1.0
streamlit==1.21.0
pillow==9.5.0